Refactored the engines package.
