+++ src/main/scala-2.11/markets/engines/Matching.scala
@@ -15,11 +15,12 @@ limitations under the License.
 */
 package markets.engines
 
+import markets.mechanisms.MatchingMechanism
 import markets.orders.{AskOrder, BidOrder}
 
 
 /** A matched pair of ask and bid orders generated by some
-  * [[markets.engines.MatchingEngine `MatchingEngine`]].
+  * [[MatchingMechanism `MatchingMechanism`]].
   *
   * @param askOrder
   * @param bidOrder
